% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-clpAPI.R
\name{addRowsCLP}
\alias{addRowsCLP}
\alias{addColsCLP}
\alias{delRowsCLP}
\alias{delColsCLP}
\title{Add and delete rows and columns, clpAPI style}
\usage{
addRowsCLP(lp, nrows, lb, ub, rowst, cols, val)

addColsCLP(lp, ncols, lb, ub, obj, colst, rows, val)

delRowsCLP(lp, num, i)

delColsCLP(lp, num, j)
}
\arguments{
\item{lp}{A \code{\linkS4class{clpPtr}} object.}

\item{nrows, ncols}{Number of rows or columns to add.}

\item{lb, ub}{Bounds for the new rows or columns.}

\item{rowst, colst}{Integer vectors of starts, length \code{n + 1}.}

\item{cols, rows}{0-based indices of the new entries.}

\item{val}{Numeric vector of the new entries.}

\item{obj}{Objective coefficients for the new columns.}

\item{num}{Number of rows or columns to delete.}

\item{i, j}{0-based positions to delete.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Add and delete rows and columns, clpAPI style
}
\examples{
lp <- initProbCLP()
loadMatrixCLP(lp, 2, 0, c(0, 0, 0), integer(0), numeric(0))
addRowsCLP(lp, 1, -1e30, 4, c(0, 2), c(0, 1), c(1, 1))
getNumRowsCLP(lp)
delProbCLP(lp)
}
