% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{clp_is_proven_optimal}
\alias{clp_is_proven_optimal}
\alias{clp_is_proven_primal_infeasible}
\alias{clp_is_proven_dual_infeasible}
\alias{clp_is_abandoned}
\alias{clp_is_primal_objective_limit_reached}
\alias{clp_is_dual_objective_limit_reached}
\alias{clp_is_iteration_limit_reached}
\alias{clp_primal_feasible}
\alias{clp_dual_feasible}
\title{Feasibility and optimality flags}
\usage{
clp_is_proven_optimal(model)

clp_is_proven_primal_infeasible(model)

clp_is_proven_dual_infeasible(model)

clp_is_abandoned(model)

clp_is_primal_objective_limit_reached(model)

clp_is_dual_objective_limit_reached(model)

clp_is_iteration_limit_reached(model)

clp_primal_feasible(model)

clp_dual_feasible(model)
}
\arguments{
\item{model}{A \code{"clp_model"} object.}
}
\value{
A single logical value.
}
\description{
Convenience predicates over the last solve, mirroring the OSI-style
queries in the Clp callable library.
}
\examples{
model <- clp_model()
clp_load_problem(model, 1, 1, c(0L, 1L), 0L, 1, obj = 1, rowlb = 1)
clp_initial_solve(model)
clp_is_proven_optimal(model)
clp_free(model)
}
