% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowlevel.R
\name{clp_num_rows}
\alias{clp_num_rows}
\alias{clp_num_cols}
\alias{clp_num_elements}
\title{Size of a model}
\usage{
clp_num_rows(model)

clp_num_cols(model)

clp_num_elements(model)
}
\arguments{
\item{model}{A \code{"clp_model"} object.}
}
\value{
The number of rows, columns or stored matrix entries.
  \code{clp_num_elements()} returns a double, since Clp counts entries in a
  type that may exceed the range of an R integer.
}
\description{
Size of a model
}
\examples{
model <- clp_model()
clp_resize(model, 3, 4)
c(clp_num_rows(model), clp_num_cols(model))
clp_free(model)
}
