% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{clp_objective_value}
\alias{clp_objective_value}
\alias{clp_objective_offset}
\alias{clp_set_objective_offset}
\title{Objective value and offset}
\usage{
clp_objective_value(model)

clp_objective_offset(model)

clp_set_objective_offset(model, value)
}
\arguments{
\item{model}{A \code{"clp_model"} object.}

\item{value}{A constant added to the objective.}
}
\value{
A number, or \code{NULL} invisibly for the setter.
}
\description{
Objective value and offset
}
\examples{
model <- clp_model()
clp_load_problem(model, 1, 1, c(0L, 1L), 0L, 1, obj = -1, rowub = 2)
clp_initial_solve(model)
clp_objective_value(model)
clp_free(model)
}
