% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{clp_options_do_dual}
\alias{clp_options_do_dual}
\alias{clp_options_set_do_dual}
\alias{clp_options_do_singleton}
\alias{clp_options_set_do_singleton}
\alias{clp_options_do_doubleton}
\alias{clp_options_set_do_doubleton}
\alias{clp_options_do_tripleton}
\alias{clp_options_set_do_tripleton}
\alias{clp_options_do_tighten}
\alias{clp_options_set_do_tighten}
\alias{clp_options_do_forcing}
\alias{clp_options_set_do_forcing}
\alias{clp_options_do_implied_free}
\alias{clp_options_set_do_implied_free}
\alias{clp_options_do_dupcol}
\alias{clp_options_set_do_dupcol}
\alias{clp_options_do_duprow}
\alias{clp_options_set_do_duprow}
\alias{clp_options_do_singleton_column}
\alias{clp_options_set_do_singleton_column}
\alias{clp_options_presolve_actions}
\alias{clp_options_set_presolve_actions}
\alias{clp_options_substitution}
\alias{clp_options_set_substitution}
\alias{clp_options_infeasible_return}
\alias{clp_options_set_infeasible_return}
\title{Individual presolve transformations}
\usage{
clp_options_do_dual(options)

clp_options_set_do_dual(options, value)

clp_options_do_singleton(options)

clp_options_set_do_singleton(options, value)

clp_options_do_doubleton(options)

clp_options_set_do_doubleton(options, value)

clp_options_do_tripleton(options)

clp_options_set_do_tripleton(options, value)

clp_options_do_tighten(options)

clp_options_set_do_tighten(options, value)

clp_options_do_forcing(options)

clp_options_set_do_forcing(options, value)

clp_options_do_implied_free(options)

clp_options_set_do_implied_free(options, value)

clp_options_do_dupcol(options)

clp_options_set_do_dupcol(options, value)

clp_options_do_duprow(options)

clp_options_set_do_duprow(options, value)

clp_options_do_singleton_column(options)

clp_options_set_do_singleton_column(options, value)

clp_options_presolve_actions(options)

clp_options_set_presolve_actions(options, value)

clp_options_substitution(options)

clp_options_set_substitution(options, value)

clp_options_infeasible_return(options)

clp_options_set_infeasible_return(options, value)
}
\arguments{
\item{options}{A \code{"clp_options"} object.}

\item{value}{The new value; 0 or 1 for the switches.}
}
\value{
The getters return an integer; the setters \code{NULL} invisibly.
}
\description{
Switch single presolve transformations on or off in a
\code{\link{clp_options}} object.  Each getter returns Clp's current
setting for that transformation.
}
\examples{
opts <- clp_options()
clp_options_set_do_dual(opts, 1L)
clp_options_do_dual(opts)
clp_options_free(opts)
}
