% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{clp_options_free}
\alias{clp_options_free}
\title{Release a Clp options object}
\usage{
clp_options_free(options)
}
\arguments{
\item{options}{A \code{"clp_options"} object.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Release a Clp options object
}
\examples{
opts <- clp_options()
clp_options_free(opts)
}
