% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowlevel.R
\name{clp_resize}
\alias{clp_resize}
\title{Change the size of a model}
\usage{
clp_resize(model, nrows, ncols)
}
\arguments{
\item{model}{A \code{"clp_model"} object.}

\item{nrows, ncols}{New numbers of rows and columns.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Change the size of a model
}
\examples{
model <- clp_model()
clp_resize(model, 2, 3)
clp_num_cols(model)
clp_free(model)
}
