% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{clp_sum_primal_infeasibilities}
\alias{clp_sum_primal_infeasibilities}
\alias{clp_sum_dual_infeasibilities}
\alias{clp_num_primal_infeasibilities}
\alias{clp_num_dual_infeasibilities}
\alias{clp_check_solution}
\title{Infeasibility measures}
\usage{
clp_sum_primal_infeasibilities(model)

clp_sum_dual_infeasibilities(model)

clp_num_primal_infeasibilities(model)

clp_num_dual_infeasibilities(model)

clp_check_solution(model)
}
\arguments{
\item{model}{A \code{"clp_model"} object.}
}
\value{
A number of violations, or the sum of their sizes.
}
\description{
Infeasibility measures
}
\examples{
model <- clp_model()
clp_load_problem(model, 1, 1, c(0L, 1L), 0L, 1, obj = 1, rowlb = 1)
clp_initial_solve(model)
clp_sum_primal_infeasibilities(model)
clp_free(model)
}
