% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowlevel.R
\name{clp_unbounded_ray}
\alias{clp_unbounded_ray}
\alias{clp_infeasibility_ray}
\title{Rays certifying unboundedness or infeasibility}
\usage{
clp_unbounded_ray(model)

clp_infeasibility_ray(model)
}
\arguments{
\item{model}{A \code{"clp_model"} object.}
}
\value{
A numeric vector, empty when Clp holds no such ray.
}
\description{
Rays certifying unboundedness or infeasibility
}
\examples{
model <- clp_model()
clp_load_problem(model, 1, 1, c(0L, 1L), 0L, 1, obj = -1, rowub = clp_inf())
clp_initial_solve(model)
clp_unbounded_ray(model)
clp_free(model)
}
