% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowlevel.R
\name{clp_version}
\alias{clp_version}
\alias{clp_features}
\title{Version of the Clp library in use}
\usage{
clp_version()

clp_features()
}
\value{
\code{clp_version()} returns a list with the version string and its
  major, minor and release components.  \code{clp_features()} returns a
  named logical vector saying which optional entry points this build of
  Clp provides.
}
\description{
Version of the Clp library in use
}
\examples{
clp_version()
clp_features()
}
