% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-clpAPI.R
\name{clpPtr-accessors}
\alias{clpPtr-accessors}
\alias{clpPointer}
\alias{clpPtrType}
\alias{clpPtrType<-}
\alias{isNULLpointerCLP}
\alias{isCLPpointer}
\title{Accessors for clpPtr objects}
\usage{
clpPointer(object)

clpPtrType(object)

clpPtrType(object) <- value

isNULLpointerCLP(object)

isCLPpointer(object)
}
\arguments{
\item{object}{A \code{\linkS4class{clpPtr}} object.}

\item{value}{A replacement value.}
}
\value{
\code{clpPointer()} the external pointer, \code{clpPtrType()} a
  string, \code{isNULLpointerCLP()} and \code{isCLPpointer()} a logical.
}
\description{
Accessors for clpPtr objects
}
\examples{
lp <- initProbCLP()
clpPtrType(lp)
delProbCLP(lp)
}
