% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-clpAPI.R
\docType{class}
\name{clpPtr-class}
\alias{clpPtr-class}
\title{A pointer to a Clp problem, clpAPI style}
\description{
An S4 class with the same shape as the \code{clpPtr} class of the archived
\pkg{clpAPI} package, so that code written against that package keeps
working.  \code{\link{initProbCLP}} creates one.
}
\section{Slots}{

\describe{
\item{\code{clpPtrType}}{A string describing the pointer, \code{"clp_prob"}.}

\item{\code{clpPointer}}{The external pointer to the Clp model.}
}}

\examples{
lp <- initProbCLP()
isCLPpointer(lp)
delProbCLP(lp)
}
