% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-clpAPI.R
\name{copyNamesCLP}
\alias{copyNamesCLP}
\alias{dropNamesCLP}
\alias{lengthNamesCLP}
\alias{probNameCLP}
\alias{setRowNameCLP}
\alias{setColNameCLP}
\alias{printModelCLP}
\alias{readMPSCLP}
\alias{writeMPSCLP}
\alias{saveModelCLP}
\alias{restoreModelCLP}
\alias{modifyCoefficientCLP}
\alias{isAvailableFuncCLP}
\title{Names and files, clpAPI style}
\usage{
copyNamesCLP(lp, cnames, rnames)

dropNamesCLP(lp)

lengthNamesCLP(lp)

probNameCLP(lp, pname)

setRowNameCLP(lp, i, rname)

setColNameCLP(lp, j, cname)

printModelCLP(lp, prefix = "CLPmodel")

readMPSCLP(lp, fname, keepNames = TRUE, ignoreErrors = FALSE)

writeMPSCLP(lp, fname, formatType = 0, numberAcross = 1, objSense = 1)

saveModelCLP(lp, fname)

restoreModelCLP(lp, fname)

modifyCoefficientCLP(lp, i, j, el, keepZero = TRUE)

isAvailableFuncCLP(funcname)
}
\arguments{
\item{lp}{A \code{\linkS4class{clpPtr}} object.}

\item{cnames, rnames}{Character vectors of column and row names.}

\item{pname}{A problem name.}

\item{i, j}{0-based row and column positions.}

\item{rname, cname}{A single name.}

\item{prefix}{A prefix for \code{printModelCLP()}.}

\item{fname}{A file name.}

\item{keepNames}{Keep the names found in an MPS file.}

\item{ignoreErrors}{Carry on after errors in an MPS file.}

\item{formatType, numberAcross, objSense}{Passed on to the MPS writer.}

\item{el}{A new matrix coefficient.}

\item{keepZero}{Keep zero entries in the sparse structure.}

\item{funcname}{Name of an optional Clp entry point.}
}
\value{
Character or integer values as in \pkg{clpAPI}; the setters return
  \code{NULL} invisibly.
}
\description{
Names and files, clpAPI style
}
\examples{
lp <- initProbCLP()
loadMatrixCLP(lp, 2, 1, c(0, 1, 2), c(0, 0), c(1, 1))
copyNamesCLP(lp, c("x", "y"), "budget")
lengthNamesCLP(lp)
delProbCLP(lp)
}
