% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-clpAPI.R
\name{getNumRowsCLP}
\alias{getNumRowsCLP}
\alias{getNumColsCLP}
\alias{getNumNnzCLP}
\alias{getObjCoefsCLP}
\alias{chgObjCoefsCLP}
\alias{getColLowerCLP}
\alias{chgColLowerCLP}
\alias{getColUpperCLP}
\alias{chgColUpperCLP}
\alias{getRowLowerCLP}
\alias{chgRowLowerCLP}
\alias{getRowUpperCLP}
\alias{chgRowUpperCLP}
\alias{getVecStartCLP}
\alias{getVecLenCLP}
\alias{getIndCLP}
\alias{getNnzCLP}
\title{Problem data, clpAPI style}
\usage{
getNumRowsCLP(lp)

getNumColsCLP(lp)

getNumNnzCLP(lp)

getObjCoefsCLP(lp)

chgObjCoefsCLP(lp, objCoef)

getColLowerCLP(lp)

chgColLowerCLP(lp, lb)

getColUpperCLP(lp)

chgColUpperCLP(lp, ub)

getRowLowerCLP(lp)

chgRowLowerCLP(lp, rlb)

getRowUpperCLP(lp)

chgRowUpperCLP(lp, rub)

getVecStartCLP(lp)

getVecLenCLP(lp)

getIndCLP(lp)

getNnzCLP(lp)
}
\arguments{
\item{lp}{A \code{\linkS4class{clpPtr}} object.}

\item{objCoef}{Objective coefficients.}

\item{lb, ub}{Column bounds.}

\item{rlb, rub}{Row bounds.}
}
\value{
The getters return numeric or integer vectors; the setters
  \code{NULL} invisibly.
}
\description{
Problem data, clpAPI style
}
\examples{
lp <- initProbCLP()
loadMatrixCLP(lp, 2, 0, c(0, 0, 0), integer(0), numeric(0))
chgObjCoefsCLP(lp, c(1, 2))
getObjCoefsCLP(lp)
delProbCLP(lp)
}
