% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-clpAPI.R
\name{getSolStatusCLP}
\alias{getSolStatusCLP}
\alias{getObjValCLP}
\alias{getColPrimCLP}
\alias{getColDualCLP}
\alias{getRowPrimCLP}
\alias{getRowDualCLP}
\alias{status_codeCLP}
\alias{return_codeCLP}
\title{Solutions and status, clpAPI style}
\usage{
getSolStatusCLP(lp)

getObjValCLP(lp)

getColPrimCLP(lp)

getColDualCLP(lp)

getRowPrimCLP(lp)

getRowDualCLP(lp)

status_codeCLP(code)

return_codeCLP(code)
}
\arguments{
\item{lp}{A \code{\linkS4class{clpPtr}} object.}

\item{code}{A status or return code.}
}
\value{
Numeric vectors for the solution accessors, an integer for
  \code{getSolStatusCLP()}, a string for the code descriptions.
}
\description{
Solutions and status, clpAPI style
}
\examples{
lp <- initProbCLP()
loadProblemCLP(lp, 1, 1, c(0, 1), 0, 1, obj_coef = -1, rlb = -1e30, rub = 2)
setLogLevelCLP(lp, 0)
solveInitialCLP(lp)
status_codeCLP(getSolStatusCLP(lp))
delProbCLP(lp)
}
