% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-clpAPI.R
\name{initProbCLP}
\alias{initProbCLP}
\alias{delProbCLP}
\title{Create and delete a problem, clpAPI style}
\usage{
initProbCLP(ptrtype = "clp_prob")

delProbCLP(lp)
}
\arguments{
\item{ptrtype}{A string stored in the returned object.}

\item{lp}{A \code{\linkS4class{clpPtr}} object.}
}
\value{
\code{initProbCLP()} returns a \code{\linkS4class{clpPtr}};
  \code{delProbCLP()} returns \code{NULL} invisibly.
}
\description{
As in \code{\link{clp_model}}, a new problem starts with Clp's message
level at 0; \code{setLogLevelCLP()} turns the solver's reporting on.
}
\examples{
lp <- initProbCLP()
setLogLevelCLP(lp, 0)
delProbCLP(lp)
}
