% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-clpAPI.R
\name{resizeCLP}
\alias{resizeCLP}
\title{Change the size of a problem, clpAPI style}
\usage{
resizeCLP(lp, nrows, ncols)
}
\arguments{
\item{lp}{A \code{\linkS4class{clpPtr}} object.}

\item{nrows, ncols}{New dimensions.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Change the size of a problem, clpAPI style
}
\examples{
lp <- initProbCLP()
resizeCLP(lp, 2, 2)
getNumRowsCLP(lp)
delProbCLP(lp)
}
