% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-clpAPI.R
\name{setObjDirCLP}
\alias{setObjDirCLP}
\alias{getObjDirCLP}
\alias{setLogLevelCLP}
\alias{getLogLevelCLP}
\alias{scaleModelCLP}
\alias{getScaleFlagCLP}
\alias{setNumberIterationsCLP}
\alias{setMaximumIterationsCLP}
\alias{getMaximumIterationsCLP}
\alias{getHitMaximumIterationsCLP}
\alias{setMaximumSecondsCLP}
\alias{getMaximumSecondsCLP}
\title{Solver settings, clpAPI style}
\usage{
setObjDirCLP(lp, lpdir)

getObjDirCLP(lp)

setLogLevelCLP(lp, amount)

getLogLevelCLP(lp)

scaleModelCLP(lp, mode)

getScaleFlagCLP(lp)

setNumberIterationsCLP(lp, iterations)

setMaximumIterationsCLP(lp, iterations)

getMaximumIterationsCLP(lp)

getHitMaximumIterationsCLP(lp)

setMaximumSecondsCLP(lp, seconds)

getMaximumSecondsCLP(lp)
}
\arguments{
\item{lp}{A \code{\linkS4class{clpPtr}} object.}

\item{lpdir}{1 to minimise, -1 to maximise.}

\item{amount}{Log level, 0 to 4.}

\item{mode}{Scaling mode.}

\item{iterations}{Iteration count or limit.}

\item{seconds}{Time limit in seconds.}
}
\value{
The getters return a number; the setters \code{NULL} invisibly.
}
\description{
Solver settings, clpAPI style
}
\examples{
lp <- initProbCLP()
setObjDirCLP(lp, -1)
getObjDirCLP(lp)
delProbCLP(lp)
}
