% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-clpAPI.R
\name{solveInitialCLP}
\alias{solveInitialCLP}
\alias{solveInitialDualCLP}
\alias{solveInitialPrimalCLP}
\alias{solveInitialBarrierCLP}
\alias{solveInitialBarrierNoCrossCLP}
\alias{primalCLP}
\alias{dualCLP}
\alias{idiotCLP}
\title{Solve a problem, clpAPI style}
\usage{
solveInitialCLP(lp)

solveInitialDualCLP(lp)

solveInitialPrimalCLP(lp)

solveInitialBarrierCLP(lp)

solveInitialBarrierNoCrossCLP(lp)

primalCLP(lp, ifValP = 0)

dualCLP(lp, ifValP = 0)

idiotCLP(lp, thd = 0)
}
\arguments{
\item{lp}{A \code{\linkS4class{clpPtr}} object.}

\item{ifValP}{Pass 1 for a values pass, 0 otherwise.}

\item{thd}{Effort level for the idiot crash.}
}
\value{
An integer return code from Clp; \code{idiotCLP()} returns
  \code{NULL} invisibly.
}
\description{
Solve a problem, clpAPI style
}
\examples{
lp <- initProbCLP()
loadProblemCLP(lp, 1, 1, c(0, 1), 0, 1, obj_coef = -1, rlb = -1e30, rub = 2)
setLogLevelCLP(lp, 0)
solveInitialCLP(lp)
getColPrimCLP(lp)
delProbCLP(lp)
}
