% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-clpAPI.R
\name{versionCLP}
\alias{versionCLP}
\title{The Clp version, clpAPI style}
\usage{
versionCLP()
}
\value{
The version of the Clp library as a string.
}
\description{
The Clp version, clpAPI style
}
\examples{
versionCLP()
}
